fix(#131): apply Copilot review fixes from PR #138
- WebSocketServer: store ScheduledFuture for heartbeat; cancel on stop()
  to prevent orphaned tasks across stop/restart cycles
- WebSocketServer: treat ::1/[::1] as loopback bindings for origin
  validation, consistent with localhost/127.0.0.1
- WebSocketStreamingListener: set outcome to null in tool_called message
  since ToolCallEvent carries no success/failure signal
- TaskCompletedMessage: change tokenCount int -> long; use -1 (unknown)
  sentinel matching TaskMetrics.totalTokens convention
- build.gradle.kts: promote agentensemble-review from compileOnly to api
  to prevent NoClassDefFoundError for consumers who reference review types
- docs/guides/live-dashboard.md: fix Origin Validation section to match
  actual implementation; update dependency comment for api promotion
- docs/getting-started/installation.md: align BOM version to 2.1.0;
  remove explicit version from agentensemble-web (managed by BOM)
- memory-bank/progress.md: fix API typo Ensemble.builder().webDashboard()
- Tests: heartbeatFutureCancelledOnStop, ipv6LoopbackBindingTreatedAsLocalBinding,
  onToolCall_outcomeIsNull, onTaskComplete_tokenCountIsMinusOne,
  taskCompletedMessage_sentinelRoundTrip, toolCalledMessage_nullOutcome

+++ agentensemble-web/src/main/java/net/agentensemble/web/protocol/TaskCompletedMessage.java
@@ -11,7 +11,8 @@ import java.time.Instant;
  * @param agentRole       role of the agent that executed the task
  * @param completedAt     when the task completed
  * @param durationMs      elapsed time in milliseconds
- * @param tokenCount      total tokens used (0 if metrics not available)
+ * @param tokenCount      total tokens used; {@code -1} when the LLM provider did not
+ *                        return usage metadata (consistent with {@code TaskMetrics.totalTokens})
  * @param toolCallCount   number of tool calls made during this task
  */
 public record TaskCompletedMessage(
@@ -21,6 +22,6 @@ public record TaskCompletedMessage(
         String agentRole,
         Instant completedAt,
         long durationMs,
-        int tokenCount,
+        long tokenCount,
         int toolCallCount)
         implements ServerMessage {}
